+++ src/extensions.ts
@@ -1,4 +1,3 @@
-import { List } from './model'
 export const keys = Object.keys as <T>(o: T) => (Extract<keyof T, string>)[]
 
 export const last = <T>(xs: ReadonlyArray<T>): T | undefined => xs[xs.length - 1]
@@ -59,6 +58,12 @@ export const hash = (str: string): number => {
   return hashValue
 }
 
+export type List<T> = ReadonlyArray<T>
+
 export const isEmpty = <T>(value: List<T> | undefined): boolean => !notEmpty(value)
 
 export const notEmpty = <T>(value: List<T> | undefined): boolean => (value?.length ?? 0) > 0
+
+export const duplicates = <T>(list: List<T>): List<T> => list.filter((element: T, i: number) => list.includes(element, i + 1))
+
+export const count = <T>(list: List<T>, condition: (element: T) => boolean): number => list.filter(condition).length