+++ src/interpreter/runtimeModel.ts
@@ -1,5 +1,5 @@
-import { Environment, is, Method, Module, Name, Node, Variable, Singleton, Expression, List, Id, Body, Assignment, Return, Reference, Self, Literal, LiteralValue, New, Send, Super, If, Try, Throw, Test, Program } from '../model'
-import { get, last } from '../extensions'
+import { Environment, is, Method, Module, Name, Node, Variable, Singleton, Expression, Id, Body, Assignment, Return, Reference, Self, Literal, LiteralValue, New, Send, Super, If, Try, Throw, Test, Program } from '../model'
+import { get, last, List } from '../extensions'
 import { v4 as uuid } from 'uuid'
 import { Interpreter } from './interpreter'
 
@@ -22,11 +22,20 @@ export type NativeFunction = (this: Evaluation, self: RuntimeObject, ...args: Ru
 // ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
 
 export class WollokReturn extends Error {
-  constructor(readonly instance?: RuntimeObject){
-    super()
-    this.name = this.constructor.name
-    this.message = 'Unhandled return on empty stack'
+  private static instance = new WollokReturn()
+
+  static of(instance?: RuntimeObject): WollokReturn {
+    WollokReturn.instance._instance = instance
+    WollokReturn.instance.name = this.name
+    WollokReturn.instance.message = 'Unhandled return on empty stack'
+    return WollokReturn.instance
   }
+
+  private _instance?: RuntimeObject
+
+  get instance(): RuntimeObject | undefined { return this._instance }
+
+  private constructor() { super() }
 }
 
 export class WollokException extends Error {
@@ -34,7 +43,7 @@ export class WollokException extends Error {
     try {
       const fullTrace: string = new Interpreter(this.evaluation).send('getStackTraceAsString', this.instance)!.innerString!
       return fullTrace.slice(fullTrace.indexOf('\n') + 1).trimEnd()
-    } catch (error) { return `Could not retrieve Wollok stack due to error: ${error.stack}` }
+    } catch (error) { return `Could not retrieve Wollok stack due to error: ${error instanceof Error ? error.stack : error}` }
   }
 
   get message(): string {
@@ -119,17 +128,20 @@ export class Frame extends Context {
     return this.node.match({
       Entity: node => `${node.fullyQualifiedName()}`,
       // TODO: Add fqn to method
-      Method: node => `${node.parent().fullyQualifiedName()}.${node.name}(${node.parameters.map(parameter => parameter.name).join(', ')})`,
+      Method: node => `${node.parent.fullyQualifiedName()}.${node.name}(${node.parameters.map(parameter => parameter.name).join(', ')})`,
       Catch: node => `catch(${node.parameter.name}: ${node.parameterType.name})`,
       Environment: () => 'root',
       Node: node => `${node.kind}`,
     })
   }
 
+  // TODO: On error report, this tells the node line, but not the actual error line.
+  //        For example, an error on a test would say the test start line, not the line where the error occurred.
   get sourceInfo(): string {
-    // TODO: Make singleton an expression and avoid the literal here
-    const sourceMap = this.node.sourceMap ?? (this.node.is('Method') && this.node.name === '<apply>' ? this.node.ancestors().find(is('Literal'))?.sourceMap : undefined)
-    return `${this.node.sourceFileName() ?? '--'}:${sourceMap ? sourceMap.start.line + ':' + sourceMap.start.column : '--'}`
+    const target = this.node.is('Method') && this.node.name === '<apply>'
+      ? this.node.parent
+      : this.node
+    return target.sourceInfo()
   }
 
   protected baseCopy(contextCache: Map<Id, Context>): Frame {
@@ -201,7 +213,7 @@ export class RuntimeObject extends Context {
   }
 
   assertIsException(): asserts this is BasicRuntimeObject<Error | undefined> {
-    if (!this.module.inherits(this.module.environment().getNodeByFQN('wollok.lang.Exception'))) throw new TypeError(`Expected an instance of Exception but got a ${this.module.fullyQualifiedName()} instead`)
+    if (!this.module.inherits(this.module.environment.getNodeByFQN('wollok.lang.Exception'))) throw new TypeError(`Expected an instance of Exception but got a ${this.module.fullyQualifiedName()} instead`)
     if(this.innerValue && !(this.innerValue instanceof Error)) throw new TypeError('Malformed Runtime Object: Exception inner value, if defined, should be an Error')
   }
 
@@ -237,7 +249,7 @@ export class Evaluation {
 
     environment.forEach(node => {
       if(node.is('Method') && node.isNative())
-        evaluation.natives.set(node, get(natives, `${node.parent()!.fullyQualifiedName()}.${node.name}`)!)
+        evaluation.natives.set(node, get(natives, `${node.parent.fullyQualifiedName()}.${node.name}`)!)
     })
 
     const globalSingletons = environment.descendants().filter((node: Node): node is Singleton => node.is('Singleton') && !!node.name)
@@ -245,14 +257,14 @@ export class Evaluation {
       evaluation.rootFrame.set(module.fullyQualifiedName(), evaluation.instantiate(module))
 
 
-    const globalConstants = environment.descendants().filter((node: Node): node is Variable => node.is('Variable') && node.parent().is('Package'))
+    const globalConstants = environment.descendants().filter((node: Node): node is Variable => node.is('Variable') && node.parent.is('Package'))
     for (const constant of globalConstants)
       evaluation.rootFrame.set(constant.fullyQualifiedName(), evaluation.exec(constant.value))
 
 
     for (const module of globalSingletons) {
       const instance = evaluation.object(module.fullyQualifiedName())
-      for (const field of module.defaultFieldValues().keys()) // TODO: Add an allFields method
+      for (const field of module.allFields())
         instance!.get(field.name)
     }
 
@@ -341,10 +353,11 @@ export class Evaluation {
       }
     } catch(error) {
       if(error instanceof WollokException || error instanceof WollokReturn) throw error
+
       const moduleFQN = error instanceof RangeError && error.message === 'Maximum call stack size exceeded'
         ? 'wollok.lang.StackOverflowException'
         : 'wollok.lang.EvaluationError'
-      const exceptionInstance = new WollokException(this, yield* this.error(moduleFQN, {}, error))
+      const exceptionInstance = new WollokException(this, yield* this.error(moduleFQN, {}, error as Error))
       throw exceptionInstance
     }
     finally { if(frame) this.frameStack.pop() }
@@ -353,8 +366,8 @@ export class Evaluation {
   protected *execTest(node: Test): Execution<void> {
     yield node
 
-    yield* this.exec(node.body, new Frame(node, node.parent().is('Describe')
-      ? yield* this.instantiate(node.parent())
+    yield* this.exec(node.body, new Frame(node, node.parent.is('Describe')
+      ? yield* this.instantiate(node.parent)
       : this.currentFrame,
     ))
   }
@@ -370,19 +383,20 @@ export class Evaluation {
 
     if(node.isNative()) {
       const native = this.natives.get(node)
-      if(!native) throw new Error(`Missing native for ${node.parent()?.fullyQualifiedName()}.${node.name}`)
+      if(!native) throw new Error(`Missing native for ${node.parent.fullyQualifiedName()}.${node.name}`)
 
       const args = node.parameters.map(parameter => this.currentFrame.get(parameter.name)!)
 
       return (yield* native.call(this, this.currentFrame.get('self')!, ...args)) ?? undefined
     } else if(node.isConcrete()) {
       try {
-        return yield* this.exec(node.body!)
+        yield* this.exec(node.body!)
+        return
       } catch(error) {
         if(error instanceof WollokReturn) return error.instance
         else throw error
       }
-    } else throw new Error(`Can't invoke abstract method ${node.parent().fullyQualifiedName()}.${node.name}/${node.parameters.length}`)
+    } else throw new Error(`Can't invoke abstract method ${node.parent.fullyQualifiedName()}.${node.name}/${node.parameters.length}`)
   }
 
   protected *execBody(node: Body): Execution<RuntimeValue> {
@@ -416,7 +430,7 @@ export class Evaluation {
   protected *execReturn(node: Return): Execution<RuntimeValue> {
     const value = node.value && (yield* this.exec(node.value))
     yield node
-    throw new WollokReturn(value)
+    throw WollokReturn.of(value)
   }
 
   protected *execReference(node: Reference<Node>): Execution<RuntimeValue> {
@@ -427,7 +441,7 @@ export class Evaluation {
     const target = node.target()!
 
     return this.currentFrame.get(
-      target.is('Module') || target.is('Variable') && target.parent().is('Package')
+      target.is('Module') || target.is('Variable') && target.parent.is('Package')
         ? target.fullyQualifiedName()
         : node.name
     )
@@ -490,7 +504,7 @@ export class Evaluation {
     const receiver = this.currentFrame.get('self')!
     const currentMethod = node.ancestors().find(is('Method'))!
     //TODO: pass just the parent (not the FQN) to lookup?
-    const method = receiver.module.lookupMethod(currentMethod.name, node.args.length, currentMethod.parent().fullyQualifiedName())
+    const method = receiver.module.lookupMethod(currentMethod.name, node.args.length, { lookupStartFQN: currentMethod.parent.fullyQualifiedName() })
 
     if (!method) return yield* this.send('messageNotUnderstood', receiver, yield* this.reify(currentMethod.name), yield* this.list(...args))
 
@@ -515,11 +529,16 @@ export class Evaluation {
     } catch(error) {
       if(!(error instanceof WollokException)) throw error
 
-      const handler = node.catches.find(catcher => error.instance.module.inherits(catcher.parameterType.target()))
+      const errorType = error.instance.module
+      const handler = node.catches.find(catcher => {
+        const handledType = catcher.parameterType.target()
+        return handledType && errorType.inherits(handledType)
+      })
 
       if(handler) {
         result = yield* this.exec(handler.body, new Frame(handler, this.currentFrame, { [handler.parameter.name]: error.instance }))
       } else throw error
+
     } finally {
       yield* this.exec(node.always, new Frame(node, this.currentFrame))
     }