+++ test/model.test.ts
@@ -1,5 +1,6 @@
 import { should } from 'chai'
-import { Class, Method, Body, Reference, ParameterizedType } from '../src/model'
+import { Class, Field, Method, Body, Reference, ParameterizedType } from '../src/model'
+import { getCache } from '../src/decorators'
 import { restore, stub } from 'sinon'
 
 should()
@@ -13,10 +14,10 @@ describe('Wollok model', () => {
       const node = new Class({ name: 'C', supertypes: [], members: [method] })
       stub(node, 'hierarchy').returns([node])
 
-      node.cache.size.should.equal(0)
+      getCache(node).size.should.equal(0)
       const response = node.lookupMethod(method.name, method.parameters.length)
       response!.should.equal(method)
-      node.cache.get(`lookupMethod(${method.name},${method.parameters.length})`).should.equal(response)
+      getCache(node).get(`lookupMethod(${method.name},${method.parameters.length})`).should.equal(response)
     })
 
     it('should prevent a second call to the same method', () => {
@@ -26,7 +27,7 @@ describe('Wollok model', () => {
       stub(node, 'hierarchy').returns([node])
 
       node.lookupMethod(method.name, method.parameters.length)
-      node.cache.set(`lookupMethod(${method.name},${method.parameters.length})`, otherMethod)
+      getCache(node).set(`lookupMethod(${method.name},${method.parameters.length})`, otherMethod)
 
       node.lookupMethod(method.name, method.parameters.length)!.should.equal(otherMethod)
     })
@@ -81,6 +82,22 @@ describe('Wollok model', () => {
         stub(c, 'hierarchy').returns([c, b])
 
         c.isAbstract().should.be.true
+      })
+
+      it('should return correct fields for subclasses', () => {
+        const constB1 = new Field({ name: 'b1', isConstant: true  })
+        const b = new Class({ name: 'B', supertypes: [], members: [constB1], id: 'c1'  })
+        const bRef = new Reference<Class>({ name: 'B', id: 'b1r'  })
+        bRef.target = () => b as any
+        const varC1 = new Field({ name: 'c1', isConstant: true  })
+        const c = new Class({ name: 'C', members: [varC1], supertypes: [new ParameterizedType({ reference: bRef })], id: 'c1' })
+        stub(b, 'fullyQualifiedName').returns('B')
+        stub(c, 'fullyQualifiedName').returns('C')
+        stub(c, 'hierarchy').returns([c, b])
+
+        c.lookupField('d1')?.should.be.not.ok
+        c.lookupField('c1')?.should.be.ok
+        c.lookupField('b1')?.should.be.ok
       })
 
       it('should return false for classes with no abstract methods', () => {